# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: API V1
  version: v1
paths:
...
  "/api/v1/products/{id}":
    get:
      summary: Retrieves a single product
      tags:
      - Products
      parameters:
      - name: id
        in: path
        format: uuid
        required: true
        schema:
          type: string
      responses:
        '200':
          description: product found
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Product"
        '404':
          description: product not found
    put:
      summary: Updates a product (Admin/Moderator only)
      tags:
      - Products
      security:
      - Bearer: []
      parameters:
      - name: id
        in: path
        format: uuid
        required: true
        schema:
          type: string
      responses:
        '200':
          description: product updated
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Product"
        '403':
          description: forbidden
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: object
                  properties:
                    name:
                      type: string
                      example: Updated Gadget Name
                    price:
                      type: number
                      example: 249.99
                    description:
                      type: string
                      example: An updated description for this gadget.
                  required: []
              required:
              - product
    delete:
      summary: Deletes a product (Admin/Moderator only)
      tags:
      - Products
      security:
      - Bearer: []
      parameters:
      - name: id
        in: path
        format: uuid
        required: true
        schema:
          type: string
      responses:
        '204':
          description: product deleted
        '403':
          description: forbidden
...
servers:
- url: http://{defaultHost}
  variables:
    defaultHost:
      default: localhost:3000
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT token for authentication. Use format: `Bearer <token>`'
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: string
          format: decimal
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - name
      - price
...